move navigation and toolbar to app
