+++ frontend/src/modules/app/App.tsx
@@ -16,7 +16,7 @@ import { FrontendDisplaySetting, ThemeVariant } from '../../store/proto/frontend
 import { getFrontendDisplaySetting } from '../../store/controller/selectors';
 import { StoreState } from '../../store/types';
 import { darkTheme, lightTheme } from '../../styles/customTheme';
-import Routes from '../navigation/Routes';
+import Routes from './navigation/Routes';
 
 const RIGHT_CLICK_BUTTON_KEY = 2;
 const TOUCH_SCREEN_BUTTON_KEY = 0;
