+++ frontend/src/modules/app/navigation/layouts/LandingPageRoute.tsx
@@ -8,7 +8,7 @@ import { makeStyles, Theme } from '@material-ui/core/styles';
 import React, { PropsWithChildren } from 'react';
 import { Route, RouteProps } from 'react-router-dom';
 import ToolBar from '../../toolbar/ToolBar';
-import UserActivity from '../../app/UserActivity';
+import UserActivity from '../../UserActivity';
 
 const useStyles = makeStyles((theme: Theme) => ({
   root: {
