+++ frontend/src/modules/app/toolbar/ToolBar.tsx
@@ -8,11 +8,11 @@ import { AppBar, Grid } from '@material-ui/core';
 import { makeStyles, Theme } from '@material-ui/core/styles';
 import React from 'react';
 import { useLocation } from 'react-router-dom';
-import ViewDropdown from '../dashboard/views/ViewDropdown';
-import ClockIcon from '../icons/ClockIcon';
-import ModesDropdown from '../modes/ModesDropdown';
+import ViewDropdown from '../../dashboard/views/ViewDropdown';
+import ClockIcon from '../../icons/ClockIcon';
+import ModesDropdown from '../../modes/ModesDropdown';
 import { DASHBOARD_ROUTE, QUICKSTART_ROUTE } from '../navigation/constants';
-import Alarms from '../alarms/Alarms';
+import Alarms from '../../alarms/Alarms';
 import StartButton, { StartButtonModalPopup } from './StartPauseButton';
 import HeaderClock from './HeaderClock';
 import PowerIndicator from './PowerIndicator';
